Update dmx-output-mapping.md
The Dmx assignment section needed some rewriting - please check my wordning. Additionally I'd suggest to use some new images as the i button has been superseded by the little cog. Furthermore these settings are only available for modules and for ASSIGNED nodes (not for unassigned nodes, not for lines).

We should unify the spelling of that ethernet dmx protocol beginning with A :) - and as https://art-net.org.uk/ refers to it as **Art-Net**, maybe this is the correct spelling? So, not Artnet or ArtNet, but Art-Net. That's also the way how it is shown in the software btw.

Finally the DMX Merge section would, to my mind, deserves an extra section as it is a separate entry in the system menu, and a more detailled description. (I believe there is also a youtube video: https://www.youtube.com/watch?v=-IiaqMAnkaI)