correction/personalities
Any reference to the Disk or Cache columns is really wrong as these refer to older personalities (Classic, Diamond 4), and every now and then a user tries this and fails.

The update procedure should mention the 'Titan Installers' thing for consoles (works only if the personality installer's file name hasn't been changed). The Tools/Folders procedure is more a fallback if the 'Titan Installers' thing doesn't work.