Update titan-commands.md
I have not yet started to format this as tables (there is an issue filed for discussion) - so far I have just tried to correct some faults. I'd say a peer review wouldn't hurt.
I have also not changed this in accordance with Tim's reformatting/renaming the select buttons. In my notation here \<Cue\> is the cue/connect hard key while {Cue list} is the handle select button/key for the respective playback. Adjust as you deem fit.

I am also considering putting the final \<Enter\> after the sequences which need it.

removed duplicate

\<Fixture\> 1 \<Through\> 10 \<And\> \<.\>2 should be removed. Technically it works (i.e. doesn't throw an error) but doesn't make much sense as it selects the master fixtures as well as the 2nd cells (insofar the description is wrong)

lots of corrections in the record syntax (essentially most of this refers to the <Cue/Connect> button)

the syntax to record ALL BUT a certain attribute group seems to be broken (I have submitted this in the beta forum)