Update cue-list-timing.md
This has occasionally caught me as well as other operators: the Enabled/Disabled toggle implies that it switches only a specific timecode but does indeed turn on/off timecode globally. While I think this should better be changed in the software (function or design) we should at least mention it in the manual.